+++ src/main/java/insane96mcp/enhancedai/modules/skeleton/ai/EARangedBowAttackGoal.java
@@ -1,55 +1,26 @@
 package insane96mcp.enhancedai.modules.skeleton.ai;
 
-import insane96mcp.enhancedai.modules.base.ai.EAAvoidEntityGoal;
+import insane96mcp.enhancedai.modules.base.ai.RangedAttackGoal;
 import insane96mcp.enhancedai.setup.Reflection;
 import net.minecraft.sounds.SoundEvents;
 import net.minecraft.world.entity.LivingEntity;
-import net.minecraft.world.entity.ai.goal.Goal;
 import net.minecraft.world.entity.monster.AbstractSkeleton;
-import net.minecraft.world.entity.monster.Monster;
-import net.minecraft.world.entity.monster.RangedAttackMob;
 import net.minecraft.world.entity.projectile.AbstractArrow;
 import net.minecraft.world.entity.projectile.ProjectileUtil;
 import net.minecraft.world.item.BowItem;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
 
-import java.util.EnumSet;
+public class EARangedBowAttackGoal extends RangedAttackGoal<AbstractSkeleton> {
 
-public class EARangedBowAttackGoal<T extends Monster & RangedAttackMob> extends Goal {
-	private final T entity;
-	private final double moveSpeedAmp;
-	private int attackCooldown;
-	private int bowChargeTicks;
-	private float inaccuracy;
-	private final float maxAttackDistance;
-	private int attackTime = -1;
-	private int seeTime;
-	private final boolean canStrafe;
-	private boolean strafingClockwise;
-	private boolean strafingBackwards;
-	private int strafingTime = -1;
+	protected int bowChargeTicks;
 
-	public EARangedBowAttackGoal(T mob, double moveSpeedAmpIn, float maxAttackDistanceIn, boolean canStrafe) {
-		this.entity = mob;
-		this.moveSpeedAmp = moveSpeedAmpIn;
-		this.maxAttackDistance = maxAttackDistanceIn * maxAttackDistanceIn;
-		this.canStrafe = canStrafe;
-		this.setFlags(EnumSet.of(Goal.Flag.LOOK));
+	public EARangedBowAttackGoal(AbstractSkeleton mob, double moveSpeedAmpIn, float maxAttackDistanceIn, boolean canStrafe) {
+		super(mob, moveSpeedAmpIn, maxAttackDistanceIn, canStrafe);
 	}
 
-	public EARangedBowAttackGoal<T> setAttackCooldown(int attackCooldownIn) {
-		this.attackCooldown = attackCooldownIn;
-		return this;
-	}
-
-	public EARangedBowAttackGoal<T> setBowChargeTicks(int bowChargeTicks) {
+	public EARangedBowAttackGoal setBowChargeTicks(int bowChargeTicks) {
 		this.bowChargeTicks = bowChargeTicks;
-		return this;
-	}
-
-	public EARangedBowAttackGoal<T> setInaccuracy(float inaccuracy) {
-		this.inaccuracy = inaccuracy;
 		return this;
 	}
 
@@ -58,135 +29,65 @@ public class EARangedBowAttackGoal<T extends Monster & RangedAttackMob> extends 
 	 * method as well.
 	 */
 	public boolean canUse() {
-		return this.entity.getTarget() != null && this.isBowInMainhand();
+		return super.canUse() && this.isBowInMainhand();
 	}
 
 	protected boolean isBowInMainhand() {
-		return this.entity.isHolding(stack -> stack.getItem() instanceof BowItem);
+		return this.mob.isHolding(stack -> stack.getItem() instanceof BowItem);
 	}
 
 	/**
 	 * Returns whether an in-progress EntityAIBase should continue executing
 	 */
 	public boolean canContinueToUse() {
-		return this.canUse() && this.isBowInMainhand();
+		return super.canContinueToUse() && this.isBowInMainhand();
 	}
 
-	/**
-	 * Execute a one shot task or start executing a continuous task
-	 */
-	public void start() {
-		super.start();
-		this.entity.setAggressive(true);
-	}
-
-	/**
-	 * Reset the task's internal state. Called when this task is interrupted by another one
-	 */
-	public void stop() {
-		super.stop();
-		this.entity.setAggressive(false);
-		this.seeTime = 0;
-		this.attackTime = -1;
-		this.entity.stopUsingItem();
-	}
-
-	/**
-	 * Keep ticking a continuous task that has already been started
-	 */
-	public void tick() {
-		LivingEntity target = this.entity.getTarget();
-		if (target == null)
-			return;
-
-		double distanceFromTarget = this.entity.distanceToSqr(target.getX(), target.getY(), target.getZ());
-		boolean canSeeTarget = this.entity.getSensing().hasLineOfSight(target);
-		boolean flag1 = this.seeTime > 0;
-		if (canSeeTarget != flag1) {
-			this.seeTime = 0;
+	@Override
+	protected void attackTick(LivingEntity target, double distanceFromTarget, boolean canSeeTarget) {
+		int i = this.mob.getTicksUsingItem();
+		if (i > 12) {
+			this.mob.getNavigation().stop();
+			this.mob.lookAt(target, 30.0F, 30.0F);
+			this.mob.getLookControl().setLookAt(target, 30.0F, 30.0F);
 		}
+		else if (this.strafingTime > -1 && this.canStrafe()) {
+			if (distanceFromTarget > (double)(this.maxAttackDistance * 0.9F)) {
+				this.strafingBackwards = false;
+			}
+			else if (distanceFromTarget < (double)(this.maxAttackDistance * 0.8F)) {
+				this.strafingBackwards = true;
+			}
 
-		if (canSeeTarget) {
-			++this.seeTime;
+			this.mob.getMoveControl().strafe(this.strafingBackwards ? -0.5F : 0.5F, this.strafingClockwise ? 0.5F : -0.5F);
 		}
-		else {
-			--this.seeTime;
+		this.mob.getLookControl().setLookAt(target, 30.0F, 30.0F);
+
+		if (this.mob.isUsingItem()) {
+			if (!canSeeTarget && this.seeTime < -60) {
+				this.mob.stopUsingItem();
+			}
+			else if (canSeeTarget) {
+				if (i >= this.bowChargeTicks) {
+					this.mob.stopUsingItem();
+					attackEntityWithRangedAttack(this.mob, target, i);
+					this.attackTime = this.attackCooldown;
+				}
+			}
 		}
-		if (distanceFromTarget > (double)this.maxAttackDistance)
-			this.entity.getNavigation().moveTo(target, this.moveSpeedAmp);
-		else {
-
-			if (distanceFromTarget >= 49d && distanceFromTarget <= (double)this.maxAttackDistance && this.seeTime >= 20 && this.canStrafe()) {
-				++this.strafingTime;
-			}
-			else {
-				this.strafingTime = -1;
-			}
-
-			if (this.strafingTime >= 20) {
-				if ((double)this.entity.getRandom().nextFloat() < 0.3D) {
-					this.strafingClockwise = !this.strafingClockwise;
-				}
-
-				if ((double)this.entity.getRandom().nextFloat() < 0.3D) {
-					this.strafingBackwards = !this.strafingBackwards;
-				}
-
-				this.strafingTime = 0;
-			}
-
-			int i = this.entity.getTicksUsingItem();
-			if (i > 12) {
-				this.entity.getNavigation().stop();
-				this.entity.lookAt(target, 30.0F, 30.0F);
-				this.entity.getLookControl().setLookAt(target, 30.0F, 30.0F);
-			}
-			else if (this.strafingTime > -1 && this.canStrafe()) {
-				if (distanceFromTarget > (double)(this.maxAttackDistance * 0.9F)) {
-					this.strafingBackwards = false;
-				}
-				else if (distanceFromTarget < (double)(this.maxAttackDistance * 0.8F)) {
-					this.strafingBackwards = true;
-				}
-
-				this.entity.getMoveControl().strafe(this.strafingBackwards ? -0.5F : 0.5F, this.strafingClockwise ? 0.5F : -0.5F);
-			}
-			this.entity.getLookControl().setLookAt(target, 30.0F, 30.0F);
-
-			if (this.entity.isUsingItem()) {
-				if (!canSeeTarget && this.seeTime < -60) {
-					this.entity.stopUsingItem();
-				}
-				else if (canSeeTarget) {
-					if (i >= this.bowChargeTicks) {
-						this.entity.stopUsingItem();
-						attackEntityWithRangedAttack(this.entity, target, i);
-						this.attackTime = this.attackCooldown;
-					}
-				}
-			}
-			else if (--this.attackTime <= 0 && this.seeTime >= -60) {
-				this.entity.startUsingItem(ProjectileUtil.getWeaponHoldingHand(this.entity, item -> item == Items.BOW));
-			}
+		else if (--this.attackTime <= 0 && this.seeTime >= -60) {
+			this.mob.startUsingItem(ProjectileUtil.getWeaponHoldingHand(this.mob, item -> item == Items.BOW));
 		}
 	}
 
-	private boolean canStrafe() {
-		return this.canStrafe && this.entity.goalSelector.getRunningGoals().noneMatch(p -> p.getGoal() instanceof EAAvoidEntityGoal);
-	}
-
-	private void attackEntityWithRangedAttack(T entity, LivingEntity target, int chargeTicks) {
+	protected void attackEntityWithRangedAttack(AbstractSkeleton entity, LivingEntity target, int chargeTicks) {
 		ItemStack itemstack = entity.getProjectile(entity.getItemInHand(ProjectileUtil.getWeaponHoldingHand(entity, item -> item == Items.BOW)));
 		double distance = entity.distanceTo(target);
 		double distanceY = target.getY() - entity.getY();
 		float f = 1; //distanceFactor / 20.0F;
 		f = (f * f + f * 2.0F) / 3.0F;
 		AbstractArrow abstractarrowentity;
-		if (entity instanceof AbstractSkeleton skeleton)
-			abstractarrowentity = Reflection.AbstractSkeleton_getArrow(skeleton, itemstack, BowItem.getPowerForTime(chargeTicks));
-		else
-			abstractarrowentity = ProjectileUtil.getMobArrow(entity, itemstack, f);
-		//abstractarrowentity.setBaseDamage(abstractarrowentity.getBaseDamage() * (distanceFactor / 20f));
+		abstractarrowentity = Reflection.AbstractSkeleton_getArrow(entity, itemstack, BowItem.getPowerForTime(chargeTicks));
 		if (entity.getMainHandItem().getItem() instanceof net.minecraft.world.item.BowItem)
 			abstractarrowentity = ((net.minecraft.world.item.BowItem)entity.getMainHandItem().getItem()).customArrow(abstractarrowentity);
 		double dirX = target.getX() - entity.getX();
@@ -200,7 +101,4 @@ public class EARangedBowAttackGoal<T extends Monster & RangedAttackMob> extends 
 		entity.level.addFreshEntity(abstractarrowentity);
 	}
 
-	public boolean requiresUpdateEveryTick() {
-		return true;
-	}
 }